Create recipes directory, lasagna page, and anchor
Completed first 3 iterations as detailed in
The Odin Project guidelines. Making these changes
allows the repo to stay organized, functionally
and in terms of best practice. The anchors allow
navigation from the index/home page to a recipe
and the recipe page has an achor to navigate back
to the index/home page.

Next, will proceed with Iteration 3: recipe page
content.