Add img lasagna, img folder, desc, ingrid, steps
Completed iteration 3, added the a free image of
the World's Best Lasagna found on google and
downloaded to the newly made img directory. Next,
added a description with place holder text,
created an unordered list for the ingridients
section with placeholder text, and lastly made an
ordered list for the steps section with place
holder text.

Moving on to Iteration 4.